# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Odysseus VM Shell API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vm/snapshot:
    post:
      operationId: snapshotVm
      tags: [vm]
      summary: Take a VM snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnapshotInput"
      responses:
        "200":
          description: Snapshot taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmActionResult"
  /vm/config:
    get:
      operationId: getVmConfig
      tags: [vm]
      summary: Get VM hardware configuration
      responses:
        "200":
          description: VM hardware configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmConfig"
    put:
      operationId: updateVmConfig
      tags: [vm]
      summary: Update VM hardware configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VmConfigUpdate"
      responses:
        "200":
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmConfig"
...
components:
  schemas:
...
    VmActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        state:
          type: string
      required:
        - success
        - message
        - state
    VmConfig:
      type: object
      properties:
        isoPath:
          type: ["string", "null"]
        diskPath:
          type: ["string", "null"]
        ramGb:
          type: integer
        cpuCores:
          type: integer
        gpuPassthrough:
          type: ["string", "null"]
        connectionMode:
          type: string
          enum: [serial, ssh]
        sshPort:
          type: integer
        sshUser:
          type: ["string", "null"]
        sshPassword:
          type: ["string", "null"]
      required:
        - ramGb
        - cpuCores
        - connectionMode
        - sshPort
    VmConfigUpdate:
      type: object
      properties:
        isoPath:
          type: ["string", "null"]
        diskPath:
          type: ["string", "null"]
        ramGb:
          type: integer
        cpuCores:
          type: integer
        gpuPassthrough:
          type: ["string", "null"]
        connectionMode:
          type: string
          enum: [serial, ssh]
        sshPort:
          type: integer
        sshUser:
          type: ["string", "null"]
        sshPassword:
          type: ["string", "null"]
    SnapshotInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...